[Refactor] move Item.tsx file to seperate in src/components/Item folder #89
