+++ app/src/lib/index.ts
@@ -6,7 +6,7 @@ export { default as TechStackList } from './components/TechStack/TechStacks/Tech
 export { default as Contact } from './components/Contact/Contact';
 export { default as Intro } from './components/Intro/Intro';
 export { default as Gallery } from './components/Gallery/Gallery';
-export { default as Item } from './components/Gallery/Item';
+export { default as Item } from './components/Item/Item';
 export { default as Card } from './components/Card/Card';
 export { default as Masonry } from './components/Masonry/Masonry';
 export { default as Image } from './components/Masonry/Image';
